Move MainScreenViewModel.kt and MainScreenViewModelFactory.kt to right package
